third_party/blink/web_tests/platform/linux-chrome/external/wpt/reporting/reporting-isolated-across-navigations.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] Reports should be sent to the correct endpoints
  assert_equals: First endpoint should receive two reports expected 2 but got 0
Harness: the test ran to completion.
